+++ .gitignore
@@ -2,6 +2,9 @@
 .gradle
 /local.properties
 /.idea
+/.idea/compiler.xml
+/.idea/gradle.xml
+/.idea/misc.xml
 /.idea/caches
 /.idea/libraries
 /.idea/modules.xml
