+++ app/src/main/java/com/example/playlistmaker/SettingsActivity.kt
@@ -4,7 +4,9 @@ import android.content.Intent
 import android.net.Uri
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.widget.Toast
 import com.example.playlistmaker.databinding.ActivitySettingsBinding
+import java.lang.Exception
 
 class SettingsActivity : AppCompatActivity() {
     private lateinit var binding: ActivitySettingsBinding
@@ -19,19 +21,27 @@ class SettingsActivity : AppCompatActivity() {
         initializeAgreementButton()
         initializeThemeSwitcher()
     }
+
     private fun initializeBackButton() {
         binding.arrowBack.setOnClickListener {
             finish()
         }
     }
+
     private fun initializeShareButton() {
         binding.shareIcon.setOnClickListener {
             val link = getString(R.string.link)
             val shareIntent = Intent(Intent.ACTION_SEND)
             shareIntent.putExtra(Intent.EXTRA_TEXT, link)
-            startActivity(shareIntent)
+            try {
+                startActivity(shareIntent)
+            } catch (ex: Exception) {
+                Toast.makeText(applicationContext, ex.toString(), Toast.LENGTH_LONG)
+                    .show()
+            }
         }
     }
+
     private fun initializeSupportButton() {
         binding.supportIcon.setOnClickListener {
             Intent(Intent.ACTION_SENDTO).apply {
@@ -43,19 +53,32 @@ class SettingsActivity : AppCompatActivity() {
                 putExtra(Intent.EXTRA_SUBJECT, subject)
                 putExtra(Intent.EXTRA_TEXT, message)
                 if (this.resolveActivity(packageManager) != null) {
-                    startActivity(this)
+                    try {
+                        startActivity(this)
+                    } catch (ex: Exception) {
+                        Toast.makeText(applicationContext, ex.toString(), Toast.LENGTH_LONG)
+                            .show()
+                    }
                 }
             }
         }
     }
+
     private fun initializeAgreementButton() {
         binding.agreementIcon.setOnClickListener {
             val url = Uri.parse(getString(R.string.uri_practicum_offer))
             val agreementIntent = Intent(Intent.ACTION_VIEW, url)
-            startActivity(agreementIntent)
+
+            try {
+                startActivity(agreementIntent)
+            } catch (ex: Exception) {
+                Toast.makeText(applicationContext, ex.toString(), Toast.LENGTH_LONG)
+                    .show()
+            }
         }
     }
-    private fun initializeThemeSwitcher(){
+
+    private fun initializeThemeSwitcher() {
         binding.themeSwitcher.isChecked = (applicationContext as App).getCurrentTheme()
         binding.themeSwitcher.setOnCheckedChangeListener { _, isChecked ->
             (applicationContext as App).switchTheme(isChecked)
